requirement.txt
python-dotenv
OpenAI
pymongo
python-dotenv
